requirements.txt
fastapi
uvicorn
adafruit-circuitpython-ahtx0
picamera2
pillow
